django_api/requirements.txt
annotated-types==0.7.0
psycopg2-binary==2.9.10
gunicorn==22.0.0
anyio==4.10.0
asgiref==3.9.1
blinker==1.9.0
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
distlib==0.4.0
distro==1.9.0
Django==5.2.5
djangorestframework==3.16.1
filelock==3.19.1
Flask==3.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.13
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.31
langchain-text-splitters==0.3.9
langsmith==0.4.17
MarkupSafe==3.0.2
nodeenv==1.9.1
openai==1.101.0
orjson==3.11.2
packaging==25.0
platformdirs==4.3.8
pre_commit==4.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
ruff==0.12.10
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
virtualenv==20.34.0
Werkzeug==3.1.3
zstandard==0.24.0
